move consumption_period to lib
