+++ front/lib/analytics/consumption_period.ts
